requirements.txt
pyairtable>=2.3.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
requests>=2.31.0

streamlit>=1.32.0

pandas>=2.2.0
cloudinary==1.44.1
google-genai==1.60.0
PyMuPDF>=1.25.0
python-docx>=1.2.0
